[Fix] #16 - 로그아웃 error handling
